Issue 2692: Let ly:make-moment accept rationals for main and grace timing
For compatibility reasons, it remains possible to write two numbers
specifying numerator and denominator instead of the rationals.  These
forms cannot be mixed, and the two-argument form is disambiguated by
the sign of the second argument: if it is positive, it can only be a
denominator and not a grace timing.