Make distributed tarball from Git file list
This removes the requirement of having a GNUmakefile
in each and every directory of the source tree; this commit
also deletes GNUmakefiles made unnecessary this way.

dist-toplevel-txt-files toplevel target is also removed,
because it is unneeded outside of dist target and it
prevents linking of toplevel generated txt docs to
distdir when these docs are already generated.

In addition, on suggestions from Han-Wen, if dist is made with a
source directory tracked by Git, then
* the source directory is required to have no uncommitted changes nor
non-ignored untracked files,
* the time stamp of all files in the tarball is set to the time stamp
of the head of the checked-out branch.