Issue 2745: clef change confuses manual key signature
The original design of the accidental code apparently had two modes
for entries: a (notename . alter) pair for entries belonging to the
key signature, and a ((octave . notename) . (alter barnum . measurepos))
construct for actually occuring accidentals affecting the key signature
locally.  The documentation, however, also specifies an
((octave . notename) . alter) form for key signatures.  The code in general
has not been quite prepared for this.  This changes at least the accessor
functions for either car or cdr of a notename entry to only look at
their part of the signature entry.