Issue 2445: Add measure counter to LilyPond
These changes implement a new grob, `MeasureCounter,' a spanner used to
number groups of successive measures.

This commit:
- defines a new event class, `MeasureCounterEvent.'
- defines a new interface, `measure-counter-interface.' This
interface contains `count-from,' the number which begins the count, and
`columns,' which contains the NonMusicalPaperColumn grobs forming
the bounds of the new spanner.
- defines the stencil with a function in scm/music-functions.scm.
- defines an engraver, `Measure_counter_engraver.'  This engraver
is written in Scheme and placed in a new file, scm/scheme-engravers.scm,
which is added to the load list in scm/lily.scm.
- sets various grob properties to default values.
- provides two commands to frame the measures to be counted:
\startMeasureCount and \stopMeasureCount.
- provides a regression test.