Issue 2872: Provide define-session and define-session-public commands
Also moves the session logic into lily.scm instead of init.ly

A session corresponds to one .ly file on the LilyPond command line;
sessions are supposed to be processed independently in one LilyPond
run.

define-session will declare a variable that has its value recorded at
the start of the first session and reinstated at the start of each
following sessions.  This is automatically the case for every variable
defined in variables in the parser module established while loading
ly/declarations-init.ly.  However, not every changeable information
handled in the .scm parts of LilyPond can easily or reasonably be
stored in the parser module.  define-session (and
define-session-public) provide a way to get the same semantics for
variables defined in the Scheme parts of LilyPond.