convert-ly: Don't update \version when no rule is applied.
Note that this is not the same as what the existing -d
--diff-version-update command does. The behavior of convert-ly is
unchanged when rules are being applied, and -d is still used in the
same way. The change concerns itself only with the case that no rule
applies, because the version of the file is already up-to-date. In
this case, it used to be that the version in the file would be set to
version of the last rule, which would sometimes mean that the version
of the file upon output is lower than the version at input. This is
what we avoid in this patch.

This fixes issue 2670.