Change \footnote user interface and behavior to match \hide/\omit etc.
This does the following changes to the footnote user interface for
footnotes in music rather than in markups:

Time-based footnote syntax:

    \footnote [mark] offset #'Grob text \default
->  \footnote [mark] offset text [ Context. ] Grob

Like with any override, you now need to specify Context if it is not
supposed to be Bottom, like with Staff.TimeSignature.

Music-based footnote syntax:

    \footnote [mark] offset text music
->  stays identical

    \footnote [mark] offset #'Grob text music
->  \single\footnote [mark] offset text Grob music

Specifying Context.Grob is possible, but Context gets ignored.