Allows optional octavation for clefs
Clef specifications like \clef "G_(8)" or
\clef "bass^[15]" are supported.

A new property clefOctavationStyle (and its cue
clef couterpart cueClefOctavationStyle) is introduced.
When set to 'default, the octavation number is displayed
as before; 'parenthesized calls for parenthesized numbers
and 'bracketed used brackets around the number.

Furthermore, a scheme formatter function is used for
displaying the ocatavation number, so can be customized
more easily.