Issue 2935: is_main_input_ should only be set by \maininput command
is_main_input reflects the state of being in the reign of the main
input file as opposed to the init file, which corresponds to a drop of
privileges.  This drop of privileges is explicitly signified with

\maininput

and should not otherwise be tampered with in order not to restrict the
init file's capabilities.  There should be no way to arrive at a lexer
from within LilyPond or Scheme without is_main_input_ set unless one
already has one such lexer to start with.