Issue 2172: Get line and column numbers right.
This uses 1-based columns on all error output, as is the standard for
GNU programs.  It also flags version errors as being for line 1
instead of line 0 since the latter confuses Emacs' compilation mode.

The same column convention is used for point-and-click column numbers
in textedit:// URIs.  In contrast, the byte offsets into a line (also
in those URIs) are retained 0-based.

For point-and-click, this yields the correct results when using the
definitions for emacs and gvim in scm/editor.scm, the editors
configured to interpret the column number.

It is to be expected that Lilypond-specific shells (like Frescobaldi
called via Okular for point-and-click) have specialized on the
previous wrong behavior and will now exhibit the kind of one-off
behavior that non-LilyPond specific programs did previously when
encountering the error messages.

I don't see a good migration strategy for those except possibly
looking for the version number of LilyPond and deciding whether to do
one-off calculations depending on that.