Issue 3023: Papersizes with landscape/portrait suffix are not recognized by -dpapersize
This factors out the papersize recognition into a function
lookup-paper-name.  It also streamlines error handling somewhat.  A
conditional checking paper-default to be non-#f could only ever be
triggered in an error situation.  Since it does not make sense to
treat the error condition of "bad paper-alist entry" and "unrecognized
paper format" differently, this has been folded into the simpler case,
namely considering the specified width and height in
set-paper-dimensions as a substitute for the valid default paper size.