canonicalise notional octave of tonic
Nothing deliberately looks at the octave part of the pitch object storing
the tonic of a key-change event; it's not really a meaningful concept.
But comparison with equal? sees the octave, so internal_event_assignment()
reckoned two key changes differing only in this meaningless field to
be different, and therefore not permitted to occur at the same time.
This triggered false warnings of "two simultaneous key-change events"
when two instruments sharing a staff differ in the transposition used
in their music source.

To fix this, always set the octave part of the tonic pitch to the
same value.  This is done when transposing, and since \key operates by
invoking transposition this is the only place that needs to canonicalise.
The canonical octave is number -1; that is, the octave obtained by a
note name with no octave suffix, and so the octave that most commonly
occurred under the non-canonicalising system.