Issue 3049: Parser outputs Lyric events for illegal note names
Well, after this patch this is what we still do (using void-music with
duration and post events seems even worse of a followup reaction,
silently dropping duration and post events not much better).

But additionally, an error is flagged.  The complaint is "have to be
in Lyric mode for lyrics" which is accurate since in Lyric mode
arbitrary strings and markups may be accepted.  At the point the error
is flagged, we don't necessarily have encountered an unquoted string,
so "bad note name" might be even more misleading.  Before this issue
originated, the parser likely bombed out with "Unexpected STRING", and
frankly that's not really better than "have to be in Lyric mode".