Issue 3082: Let ChordNameVoice use the same performers as Voice
In particular, ties should be heeded when creating Midi from a
ChordNames context as they result in the omission of the engraved
chord name if an identical chord follows.  Reasoning for beams and
slurs is fuzzier, but an articulation different from the same music in
a regular Voice context does not seem to make sense.