Issue 3093: Declare absolute-volume-alist and instrument-equalizer-alist as session-local
At least instrument-equalizer-alist is advertised as sort of a
user-settable variable, and while we ask users to tamper with it, we
might as well restore it at the beginning of next session.  This does
not really replace a proper user interface instead of the pathetic
excuse we have for it now.