Issue 3095: \tuplet and \tupletSpan to replace \times and tupletSpannerDuration
\tuplet:
Scale the given @var{music} to tuplets.  @var{ratio} is a
fraction that specifies how many notes are played in place of the
nominal value: it will be @samp{3/2} for triplets, namely three notes
being played in place of two.  If the optional duration
@var{tuplet-span} is specified, it is used instead of
@code{tupletSpannerDuration} for grouping the tuplets.
For example,
@example
\tuplet 3/2 4 @{ c8 c c c c c @}
@end example
will result in two groups of three tuplets, each group lasting for a
quarter note.

tupletSpan:
Set @code{tupletSpannerDuration}, the length into which
@code{\tuplet} without an explicit @samp{tuplet-span} argument of its
own will group its tuplets, to the duration @var{tuplet-span}.  To
revert to the default of not subdividing the contents of a @code{\tuplet}
command without explicit @samp{tuplet-span}, use
@example
\tupletSpan \default
@end example