lilypond-book: Textwidth detection with included file (issue 3136).
Fix the automatic textwidth detection performed by lilypond-book.
This failed for lytex files that include an external file in the
preamble, when such file was compiled in an --output directory
different from the working directory, because then the included
file was missing. We use TEXINPUTS to let the latex process know
where to look for input files.