Issue 3140: Let find_create_context create Score context for Timing
More exactly: if find_create_context finds that no score context
exists yet, but creating one would make it find an alias of it, it
does so.

This is in order to let \set Timing.xxx = yyy work even in the rare
case that the Score context with its hardwired Timing alias does not
yet exist.