Issue 3153: Let music inside of #{ ... #} originate from @code{location} if set
In particular with regard to point-and-click, it has been a reoccuring
complaint that music originating from #{ ... #} has the #{ ... #}
environment as its origin.  This patch will make such music originate
from @code{location} if set to a valid input location.

It would seem somewhat peculiar that #{ ... #} just grabs whatever
meaning @code{location} happens to have at the current lexical level
and runs with it.  However, the same is already done for
@code{parser}, so it is not really out of line.