Issue 754: don't transpose generic property-setting music commands
The actual issue was that
\transpose c e { \transposition bes ... }

created a Midi corresponding to an instrument using a transposition of
ges, leaving the Midi unchanged.  Making the generic property-setting
commands (\set and \override) impervious to transposition will also
keep \transpose from tampering with user-set values.

This is particularly important since the pitch data type in LilyPond
is also being used for signifying intervals or pitch differences
rather than absolute pitches.