Issue 3201: Don't initialize dim_cache_ array via constructor syntax
This is not valid C++ and not even a documented GCC extension.  It is
probabably an oversight that GCC appears to be able to interpret this
in some manner.

Also creates a custom copy constructor for Dimension_cache for the
sole purpose of being able to initialize a Dimension_cache array
element-wise from a Dimension_cache array in the copy constructor of
Grob.  C++ clearly is lacking in some departments.