Redirect from `/website/` to `/`
Some pages on `lilypond.org` lie in the subdirectory `/website/`.  Apache's
`mod_rewrite` is being used to make these pages also available when omitting
`/website/` in the URI.  Therefore several pages can be reached from two
different URIs and are both indexed by search engines.

This commit sets up a permanent redirect from the subdirectory `/website/`
to the root of the server in order the mark the URI without `/website/` as
the preferable form.

[This commit was taken from issue #3191.]