Issue 1334: A \score-lines markup list command for multi-lines embedded scores
Like the \score markup, this is not usually called by users directly
as it would require something awkward along the lines of

\score-lines ##{ \score { ... } #}

to get this through.  Instead, using \score in contexts where only
markup lists are allowed will pass the action to the \score-lines
markup list command.

Note that there are as of yet no user-level commands spacing a markup
list in a way similar to the spacing inside of a \score markup.