Issue 3225: Decouple \defaultchild from \accepts list in contexts
The definition of a Bottom context previously was a context not
accepting any subcontexts.  Now it is a context without a
\defaultchild.  The defaultchild of a context previously was the first
found in the \accepts list (if necessary, moving it there).  While
\defaultchild was tracked in context definitions, it was not explicit
in instantiated contexts.

Decoupling those makes for more flexible arrangements of contexts.
For example, one might let Voice accept a SubVoice context without
forcing SubVoice to be created when a Bottom context is called for
when in Voice, simply by not giving Voice a \defaultchild.