Issue 3229: Prefer \relative { ... } over \relative x'' { ... }
As a followup to the commits for issue 3245, this conversion rule
replaces uses of \relative with reference pitch by those without
reference pitch for a number of patterns where convert-ly can detect the
first actual note reliably.

The rule is prepared to skip over files containing \include or \language
with a non-c-using notename language since the patterns do not work for
them.

However, this part of the rule is disabled in this commit as it turns
out that the LilyPond code base converts well (and more thoroughly)
without it.