Simplify a function pointer cast that GUB chokes on
It is conceivable that the C++ compiler has problems converting a
pointer to a C function (declared in a C language header) to a C++
function pointer.  Or that it has a compiler bug making it choke on
too complex casts.  This commit changes the cast

(SCM (*)()) apply_unpure_pure

to

(scm_t_subr) apply_unpure_pure

which hopefully addresses both of these potential candidates for GUB
compilation failure.