Issue 3237: Harmonize point-and-click information for #xxx and $xxx
Both #xxx and $xxx now retain any preexisting origin information when
producing a music expression.  In contrast, \xxx continues receiving
the origin information pertaining to the current location even if
previous point-and-click information exists.  This is consistent with
the point-and-click information for music functions as a whole
corresponding to the point of call, while the individual arguments, if
separate music expressions, retain any preexisting origin.