Issue 3245: Make \relative { ... } interpret the first pitch as an absolute one
The previous default of \relative { ... } being \relative c' { ... }
made c' more special than warranted.  Having to write
\relative f { ... } for this functionality seems awkward since f
serves no musical function but just happens to be the middle of the
scale.  Iff the scale is the standard scale.  The new definition of
\relative takes care to determine the "middle of the scale" reference
point properly even in the case of non-standard scales.