Issue 3244: Remove -d old-relative compatibility option
This removes the lily_1_8_relative variable and the code surrounding
it.  lily_1_8_relative_used apparently was unused; Scheme code for
interpreting its setting was not connected to the actual variable.

The callback for repeat-music was not actually being used and has been
removed.  While the command line option was still documented, it was
ineffective: only reenabling it in the file itself lead to any effect.

ly:relative-music::relative-callback and
ly:relative-music::no-relative-callback are now functionally
equivalent.