Don't report a programming error when aligned grob has empty extent.
(Issue 3259)

If a grob's extent is empty (undefined), alignment procedures from
Self_alignment_interface don't have any information about grob's
dimensions, so they cannot calculate the offset required to achieve
specified alignment of such grob.

However, this isn't actually a problem: if a grob's extent is empty,
this usually means that the grob itself is empty, and in that case
there is nothing to align at all.  Therefore, no programming error
should be reported.

For example, a user may remove some grob from output by overriding
its stencil to #f.  This will result in an empty extent, and Lily
shouldn't complain about being unable to align such grob, because
there is nothing to align.

Nevertheless, if the extent is empty but the stencil itself isn't
empty, the situation looks suspicious, so we issue a warning.