Issue 3274: Restructure parsing of reverts to avoid ambiguities in relation to BACKUP
A music expression could start with BACKUP (an artificially inserted
token never present in the source code itself) due to the previous
rules.  If the BACKUP mechanism is needed in different unrelated rules
applicable in similar contexts, this can lead to shift/reduce
conflicts.

While the actual (separate) code prompting this issue was rewritten in
a manner reining its BACKUP tokens in on its own and thus avoiding the
conflict triggering this issue, it still makes sense to confine the
range where BACKUP tokens are recognized to the actual limited
contexts where they are being produced, in order to keep such
conflicts down, and to stop the parser from creative reinterpretation
of situations that can only arise by programmer error.