Uses special X alignment for instrument names.
Previously, instrument names used the side-position-interface for
calculations in its alignment function.  This made no sense, as it never
had any side-support elements, requiring a kludge to get them to
align properly.

Here, we remove that kludge and align them by aligning them to their
right end and adding padding.  We also change the name of the grob array
in which vertical axis groups are stored to
system-start-text-align-elements so that the grob is not accidentally
treated as an axis-group implementing an elements array.