use X-parents to align MMRNumbers, MMRTexts and PercentRepeatCounters
It makes more sense to use X-parents than Y-parents for horizontal
alignment. To be able to do this, we have to ensure that X-parents of
MultiMeasureRestNumber, MultiMeasureRestText and PercentRepeatCounter
won't be overwritten by set_bound (these grobs are spanners).  We can
do this because parents of these grobs are spanners as well.

Expected changes in output: none.