Rename OctavateEight to ClefModifier
Name "OctavateEight" didn't make any sense.  The grob doesn't have
to be an eight, and it doesn't describe octavation, but arbitrary
transposition (not only whole octaves).  Also, the name should reflect
its affiliation with Clef grob.

The name "ClefTransposition" was also considered, but it was decided
that it would be too confusing together with clefTransposition, which
is a context property.

Also change related context properties.