Issue 3313: Add the command \offset
The ability to offset values of various properties would be a useful
enhancement of LilyPond.  Currently, this is possible for default values
of the property 'control-points using the \shape command.  The following
patch seeks to generalize the application of offsets to grob properties.
Both overrides and tweaks are supported, on the model of \shape and
\alterBroken.

Offsets are currently limited to three data types: number, number-pair,
and number-pair-list (this last is defined by this patch and represents
the type used, for example, by 'control-points).  Offsets are limited to
the properties listed in the grob descriptions contained in
`scm/define-grobs.scm'.  Offsets will be reckoned against an override
currently in effect; otherwise, the default setting from the grob
description will be used.