Rewrite Stencil::add_at_edge to treat "spacing" stencils differently
A spacing stencil is one that is non-empty in its own axis, but empty
in the orthogonal axis.  Since they don't actually have a bounding box
to convey, they don't need padding, and they don't suffer from the
equality of the lower left bounding box corner with the reference
point.