Fix PNG images.
Since libpng 1.6.0, the library rejects a class of malformed PNGs, aborting
with

  IDAT: invalid distance too far back

To be more precise, the library now relies on certain header fields in the
PNG image which have been ignored previously.  However, a
yet-to-be-identified program or library produces invalid PNGs; one example
of such PNGs is the file `lily-ledger.png'.

libpng 1.6.3 comes with a tool called `png-fix-too-far-back' to correct such
errors in PNG files (together with some other minor optimizations); this
commit contains the results of such a run.